add stock share card and connect to state
